backend/requirements.txt
asgiref==3.8.1
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django>=5.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
firebase-admin==6.4.0
gcloud==0.18.3
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
joblib>=1.2.0
jwcrypto==1.5.6
msgpack==1.1.0
numpy>=1.24.0
oauth2client==4.1.3
pandas>=2.0.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.22.0
PyJWT>=2.6.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jwt==4.1.0
python-jose>=3.3.0
pytz==2025.2
requests>=2.18.0,<2.30.0
requests-toolbelt==0.10.1
rsa==4.9
scikit-learn>=1.2.0
scipy==1.15.2
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==1.26.15
